Refactor into pages
